--- apps/docs/src/docs/composables/useTooltipController.md
@@ -1,5 +1,0 @@
-<ComposableHeader path="useTooltipController/index.ts" title="useTooltipController" />
-
-<script setup lang="ts">
-import ComposableHeader from './ComposableHeader.vue'
-</script>
