+++ packages/bootstrap-vue-next/src/components/BPopover/BPopoverOrchestrator.vue
@@ -5,14 +5,24 @@
     :key="self"
     v-bind="popover.props"
     :model-value="popover.props._modelValue"
-    :reference="popover.props._reference"
-    @update:model-value="tools.set?.(self, {...popover.props, _modelValue: $event})"
+    :target="popover.props._target"
+    @update:model-value="tools.setPopover?.(self, {...popover.props, _modelValue: $event})"
+  />
+  <component
+    :is="tooltip.component ?? BTooltip"
+    v-for="[self, tooltip] in tools.tooltips?.value"
+    :key="self"
+    v-bind="tooltip.props"
+    :model-value="tooltip.props._modelValue"
+    :target="tooltip.props._target"
+    @update:model-value="tools.setTooltip?.(self, {...tooltip.props, _modelValue: $event})"
   />
 </template>
 
 <script setup lang="ts">
 import {usePopoverController} from '../../composables/usePopoverController'
 import BPopover from './BPopover.vue'
+import BTooltip from '../BTooltip/BTooltip.vue'
 
 const tools = usePopoverController()
