--- packages/bootstrap-vue-next/src/components/BTooltip/BTooltipOrchestrator.vue
@@ -1,22 +1,0 @@
-<template>
-  <component
-    :is="tooltip.component ?? BTooltip"
-    v-for="[self, tooltip] in tools.tooltips?.value"
-    :key="self"
-    v-bind="tooltip.props"
-    :model-value="tooltip.props._modelValue"
-    :reference="tooltip.props._reference"
-    @update:model-value="tools.set?.(self, {...tooltip.props, _modelValue: $event})"
-  />
-</template>
-
-<script setup lang="ts">
-import {useTooltipController} from '../../composables/useTooltipController'
-import BTooltip from './BTooltip.vue'
-
-const tools = useTooltipController()
-
-defineExpose({
-  ...tools,
-})
-</script>
