+++ packages/bootstrap-vue-next/src/components/BTooltip/index.ts
@@ -1,2 +1,1 @@
 export {default as BTooltip} from './BTooltip.vue'
-export {default as BTooltipOrchestrator} from './BTooltipOrchestrator.vue'
