+++ packages/bootstrap-vue-next/src/composables/index.ts
@@ -4,5 +4,4 @@ export {useModal} from './useModal'
 export {useModalController} from './useModalController'
 export {useScrollspy} from './useScrollspy'
 export {useToastController} from './useToastController'
-export {useTooltipController} from './useTooltipController'
 export {usePopoverController} from './usePopoverController'
