--- packages/bootstrap-vue-next/src/composables/useTooltipController/index.ts
@@ -1,4 +1,0 @@
-import {inject} from 'vue'
-import {tooltipPluginKey} from '../../utils/keys'
-
-export const useTooltipController = () => ({...inject(tooltipPluginKey)})
